Move naming helpers to helpers/naming.ts
Relocate duplicateFeatureName, uniqueFolderName, textFolderBaseName,
createTextFeatureAt, duplicateClampName, and duplicateTabName to a
dedicated home module. Slices now import these directly.
