+++ src/store/helpers/naming.ts
@@ -1,0 +1,114 @@
+/**
+ * Copyright 2026 Franja (Frank) Povazanj
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { isProfileDegenerate, uniqueName } from '../../import'
+import {
+  generateTextShapes,
+  getTextFrameProfile,
+  type TextToolConfig,
+} from '../../text'
+import type { TextFeatureData } from '../../types/project'
+import type { Point, Project, SketchFeature, Clamp, Tab, FeatureFolder, FeatureOperation } from '../../types/project'
+import { nextUniqueGeneratedId } from './ids'
+import { normalizeFeatureZRange } from './normalize'
+
+export function duplicateFeatureName(name: string, features: SketchFeature[], totalCount: number, step: number): string {
+  if (totalCount === 1) {
+    const baseName = `${name} Copy`
+    if (!features.some((f) => f.name === baseName)) return baseName
+    let index = 2
+    while (features.some((f) => f.name === `${baseName} ${index}`)) index += 1
+    return `${baseName} ${index}`
+  }
+  let index = step
+  while (features.some((f) => f.name === `${name} Copy ${index}`)) index += 1
+  return `${name} Copy ${index}`
+}
+
+export function uniqueFolderName(preferred: string, folders: FeatureFolder[]): string {
+  return uniqueName(preferred, folders.map((folder) => folder.name))
+}
+
+export function textFolderBaseName(text: string): string {
+  const normalized = text.replace(/\s+/g, ' ').trim()
+  if (!normalized) {
+    return 'Text'
+  }
+  return normalized.length > 24 ? `${normalized.slice(0, 24)}…` : normalized
+}
+
+export function createTextFeatureAt(project: Project, config: TextToolConfig, anchor: Point): SketchFeature | null {
+  const generatedShapes = generateTextShapes(config, { x: 0, y: 0 }).filter((shape) => !isProfileDegenerate(shape.profile))
+  if (generatedShapes.length === 0) {
+    return null
+  }
+
+  const featureName = uniqueName(textFolderBaseName(config.text), project.features.map((feature) => feature.name))
+  const isFirstMachiningFeature = !project.features.some((feature) => feature.operation !== 'region')
+  const textData: TextFeatureData = {
+    text: config.text,
+    style: config.style,
+    fontId: config.fontId,
+    size: config.size,
+  }
+
+  const op: FeatureOperation = isFirstMachiningFeature ? 'add' : config.operation
+  return normalizeFeatureZRange({
+    id: nextUniqueGeneratedId(project, 'f'),
+    name: featureName,
+    kind: 'text',
+    text: textData,
+    folderId: null,
+    sketch: {
+      profile: getTextFrameProfile(config, anchor),
+      origin: { x: 0, y: 0 },
+      orientationAngle: 90,
+      dimensions: [],
+      constraints: [],
+    },
+    operation: op,
+    z_top: project.stock.thickness,
+    z_bottom: 0,
+    visible: true,
+    locked: false,
+  })
+}
+
+export function duplicateClampName(name: string, clamps: Clamp[]): string {
+  const baseName = `${name} Copy`
+  if (!clamps.some((clamp) => clamp.name === baseName)) {
+    return baseName
+  }
+
+  let index = 2
+  while (clamps.some((clamp) => clamp.name === `${baseName} ${index}`)) {
+    index += 1
+  }
+  return `${baseName} ${index}`
+}
+
+export function duplicateTabName(name: string, tabs: Tab[]): string {
+  const baseName = `${name} Copy`
+  if (!tabs.some((tab) => tab.name === baseName)) {
+    return baseName
+  }
+
+  let index = 2
+  while (tabs.some((tab) => tab.name === `${baseName} ${index}`)) {
+    index += 1
+  }
+  return `${baseName} ${index}`
+}
